+++ python/src/ouster/client/__init__.py
@@ -11,6 +11,7 @@ from ._client import NMEABaudRate
 from ._client import SensorConfig
 from ._client import get_config
 from ._client import set_config
+from ._client import LidarScan
 
 from .data import BufferT
 from .data import Packet
@@ -18,7 +19,6 @@ from .data import ImuPacket
 from .data import LidarPacket
 from .data import ChanField
 from .data import ColHeader
-from .data import LidarScan
 from .data import XYZLut
 from .data import destagger